[rayci] split config file into smaller files
with default configs in a file separate from the config definition.
