+++ raycicmd/config.go
@@ -168,172 +168,7 @@ func localDefaultConfig(envs Envs) *config {
 	}
 }
 
-// builtin ray buildkite pipeline IDs.
-const (
-	// v1 pipelines
-	rayBranchPipeline = "0183465b-c6fb-479b-8577-4cfd743b545d"
-	rayPRPipeline     = "0183465f-a222-467a-b122-3b9ea3e68094"
-
-	// v2 pipelines
-	rayV2PostmergePipeline  = "0189e759-8c96-4302-b6b5-b4274406bf89"
-	rayV2PremergePipeline   = "0189942e-0876-4b8f-80a4-617f988ec59b"
-	rayV2MicrocheckPipeline = "018f4f1e-1b73-4906-9802-92422e3badaa"
-
-	// dev only
-	rayDevPipeline = "5b097a97-ad35-4443-9552-f5c413ead11c"
-
-	// pipeline for this repo
-	rayCIPipeline = "01890992-4f1a-4cc4-b99d-f2da360eb3ab"
-)
-
-const (
-	rayCIECR = "029272617770.dkr.ecr.us-west-2.amazonaws.com"
-
-	rayBazelBuildCache = "https://bazel-cache-dev.s3.us-west-2.amazonaws.com"
-
-	defaultForgePrefix = "cr.ray.io/rayproject/"
-)
-
-var branchPipelineConfig = &config{
-	name: "ray-branch",
-
-	ArtifactsBucket: "ray-ci-artifact-branch-public",
-
-	CITemp:      "s3://ray-ci-artifact-branch-public/ci-temp/",
-	CIWorkRepo:  rayCIECR + "/rayproject/citemp",
-	ForgePrefix: defaultForgePrefix,
-
-	BuilderQueues: map[string]string{
-		"builder":         "builder_queue_branch",
-		"builder-arm64":   "builder_queue_arm64_branch",
-		"builder-windows": "builder_queue_windows_branch",
-	},
-
-	RunnerQueues: map[string]string{
-		"default":        "runner_queue_small_branch",
-		"small":          "runner_queue_small_branch",
-		"medium":         "runner_queue_medium_branch",
-		"large":          "runner_queue_branch",
-		"gpu":            "gpu_runner_queue_branch",
-		"gpu-large":      "gpu_large_runner_queue_branch",
-		"trainium":       "trainium_runner_queue_branch",
-		"windows":        "windows_queue_branch",
-		"macos":          "macos-branch",
-		"macos-arm64":    "macos-branch-arm64",
-		"medium-arm64":   "runner_queue_arm64_medium_branch",
-		"release":        "release_queue_small",
-		"release-medium": "release_queue_medium",
-	},
-
-	Env: map[string]string{
-		"BUILDKITE_BAZEL_CACHE_URL": rayBazelBuildCache,
-	},
-
-	BuildEnvKeys: []string{
-		"RAYCI_SCHEDULE",
-		"RAYCI_BISECT_TEST_TARGET",
-	},
-	HookEnvKeys: []string{"RAYCI_CHECKOUT_DIR"},
-
-	SkipTags: []string{"disabled"},
-}
-
-func prPipelineConfig(
-	name string,
-	extraEnv map[string]string,
-	maxParralelism int,
-) *config {
-	config := &config{
-		name: name,
-
-		ArtifactsBucket: "ray-ci-artifact-pr-public",
-
-		CITemp:      "s3://ray-ci-artifact-pr-public/ci-temp/",
-		CIWorkRepo:  rayCIECR + "/rayproject/citemp",
-		ForgePrefix: defaultForgePrefix,
-
-		BuilderQueues: map[string]string{
-			"builder":         "builder_queue_pr",
-			"builder-arm64":   "builder_queue_arm64_pr",
-			"builder-windows": "builder_queue_windows_pr",
-		},
-
-		RunnerQueues: map[string]string{
-			"default":     "runner_queue_small_pr",
-			"small":       "runner_queue_small_pr",
-			"medium":      "runner_queue_medium_pr",
-			"large":       "runner_queue_pr",
-			"gpu":         "gpu_runner_queue_pr",
-			"gpu-large":   "gpu_large_runner_queue_pr",
-			"trainium":    "trainium_runner_queue_pr",
-			"windows":     "windows_queue_pr",
-			"macos":       "macos",
-			"macos-arm64": "macos-pr-arm64",
-
-			"medium-arm64": "runner_queue_arm64_medium_pr",
-		},
-
-		BuilderPriority: 1,
-		RunnerPriority:  1,
-
-		Env: map[string]string{
-			"BUILDKITE_BAZEL_CACHE_URL": rayBazelBuildCache,
-			"BUILDKITE_CACHE_READONLY":  "true",
-		},
-
-		HookEnvKeys: []string{"RAYCI_CHECKOUT_DIR"},
-
-		TagFilterCommand: []string{"./ci/ci_tags_from_change.sh"},
-
-		SkipTags: []string{"disabled", "skip-on-premerge"},
-
-		ConcurrencyGroupPrefixes: []string{},
-	}
-	for k, v := range extraEnv {
-		config.Env[k] = v
-	}
-	if maxParralelism > 0 {
-		config.MaxParallelism = maxParralelism
-	}
-	return config
-}
-
-func ciDefaultConfig(envs Envs) *config {
-	pipelineID := getEnv(envs, "BUILDKITE_PIPELINE_ID")
-	switch pipelineID {
-	case rayBranchPipeline, rayV2PostmergePipeline, rayCIPipeline:
-		return branchPipelineConfig
-	case rayPRPipeline, rayV2PremergePipeline, rayDevPipeline:
-		return prPipelineConfig("ray-pr", nil, -1)
-	case rayV2MicrocheckPipeline:
-		c := prPipelineConfig(
-			"ray-pr-microcheck",
-			map[string]string{"RAYCI_MICROCHECK_RUN": "1"},
-			1,
-		)
-		c.NotifyOwnerOnFailure = true
-		c.SkipTags = append(c.SkipTags, "skip-on-microcheck")
-
-		return c
-	}
-
-	// By default, assume it is less privileged.
-	return prPipelineConfig("ray-pr", nil, -1)
-}
-
-func defaultConfig(envs Envs) *config {
-	envCI := getEnv(envs, "CI")
-	if envCI == "true" || envCI == "1" {
-		return ciDefaultConfig(envs)
-	}
-	return localDefaultConfig(envs)
-}
-
-func loadConfig(f string, envs Envs) (*config, error) {
-	if f == "" {
-		return defaultConfig(envs), nil
-	}
-
+func loadConfigFromFile(f string) (*config, error) {
 	bs, err := os.ReadFile(f)
 	if err != nil {
 		return nil, fmt.Errorf("read config: %w", err)
